bot/requirements.txt
aiogram==3.0.0
sqlalchemy==2.0.20
asyncpg==0.28.0
python-dotenv==1.0.0